+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added <podcast:socialInteract> support ([#19](https://github.com/vnali/craft-studio/discussions/19)).
 - Added <podcast:txt> support ([#20](https://github.com/vnali/craft-studio/discussions/20)).
 - Added <podcast:guid> support ([#21](https://github.com/vnali/craft-studio/discussions/21)).
+- Added <podcast:value> support ([#22](https://github.com/vnali/craft-studio/discussions/22)).
 
 ## 0.13.0 - 2023-07-19